% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFisher.R
\name{calcFisher}
\alias{calcFisher}
\title{Apply Fishers exact test}
\usage{
calcFisher(data, cluster, bin = TRUE)
}
\arguments{
\item{data}{previously created CN matrix}

\item{cluster}{group assignent, same order as data columns}

\item{bin}{Aggregate Fishers p.values to < 0.001, <0.01, 
<0.05, <0.1, >0.1}
}
\value{
vector with all p values
}
\description{
Calculates p values (Fishers exact test) for gain/loss 
and given group assignment
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

