% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBins.R
\name{createBins}
\alias{createBins}
\title{Bin CN data}
\usage{
createBins(data, ctrl, ctrlAll, statistic = "wilcoxon", binsize = 2e+05,
  output = "diff", arrayType = "auto")
}
\arguments{
\item{data}{sample data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{statistic}{statistics test to compare groups (controls 
vs smp); "t.test" or "wilcoxon" (Mann-Whitney-U Test)}

\item{binsize}{binsize}

\item{output}{"ratio" (Sample/Ctrl) or "diff" (Sample-Ctrl)}

\item{arrayType}{"auto","450k", "EPIC"; auto -> tries to automatically determine 
the array type (450k, EPIC)}
}
\value{
bins with their corresponding value
}
\description{
Calculate mean, median, sd for each bin for a given binsize.
}
\examples{
data <- data.frame(
smp1=c(-8.12, -5.225, -3.24, -3.61),
smp2=c(-5.0, -3.98, -4.06, -4.5),
smp3=c(NA, -2.48, -2.27, -2.1)
)
ctrlAll <- data.frame(
ctl1=c(1.0, -3.6, 0.7, -0.73),
ctl2=c(-0.4, -4.1, -4.2, -3.9),
ctl2=c(0.74, -1.12, -2.8, -1.67)
)
ctrl <- apply(ctrlAll, 1, "median")
createBins(data,ctrlAll, ctrl, 50000)
}

