% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSegmentMatrix.R
\name{createSegmentMatrix}
\alias{createSegmentMatrix}
\title{Create Segment-Matrix}
\usage{
createSegmentMatrix(data, p.select = 0.05)
}
\arguments{
\item{data}{input data}

\item{p.select}{0.05 selection of candidates which have a p.value
< p.select for at least 1 sample}
}
\value{
all aligned segments
}
\description{
Create Segment Matrix from Segment-list (created with \code{findSegments()})
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

