% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineArrayType.R
\name{determineArrayType}
\alias{determineArrayType}
\title{Determines Array Type}
\usage{
determineArrayType(data)
}
\arguments{
\item{data}{object, for which the array type should 
be determined}
}
\value{
"EPIC", "450k" or NULL if array type could not 
be determined
}
\description{
Tries to determine the array type (EPIC, 450k)
}
\examples{
determineArrayType(minfiData::MsetEx)
}
