% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyCutoffs.R
\name{findCutoffs}
\alias{findCutoffs}
\title{Finding cutoffs}
\usage{
findCutoffs(data, zeichne = FALSE, ignoreNAs = FALSE, eps = 0.01,
    proximity = c(0, 0))
}
\arguments{
\item{data}{Inputdata}

\item{zeichne}{Debugging}

\item{ignoreNAs}{F / T: if NAs occur, proceed, otherwise fail}

\item{eps}{difference between x points (internal calculation)}

\item{proximity}{how many previous / following rows should be 
considered for cutoff determination? c(prev, folw)}
}
\value{
identified cutoffs
}
\description{
Analyzes row-wise histograms by selecting the global maximum (0),
as well as one further local maximum right (+) and left (-).
Can make use of previous/following rows.
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

