% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyCutoffs.R
\name{findCutoffsRun}
\alias{findCutoffsRun}
\title{Finding cutoffs}
\usage{
findCutoffsRun(data, zeichne, ignoreNAs, eps, proximity)
}
\arguments{
\item{data}{Inputdata}

\item{zeichne}{Debugging}

\item{ignoreNAs}{F / T: if NAs occure, proceed, otherwise fail}

\item{eps}{difference between x points (internal calculation)}

\item{proximity}{how many previous / following rows should be 
considered for cutoff determination? c(prev, folw)}
}
\value{
identified cutoffs
}
\description{
Analyzes row-wise histograms by selecting the global maximum (0),
as well as one further local maximum right (+) and left (-)
}

