% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyCutoffs.R
\name{findCutoffsRunFast}
\alias{findCutoffsRunFast}
\title{Identifying cutoffs}
\usage{
findCutoffsRunFast(data, ignoreNAs, eps)
}
\arguments{
\item{data}{Inputdata}

\item{ignoreNAs}{F / T: if NAs occure, proceed, otherwise fail}

\item{eps}{difference between x points (internal calculation)}
}
\value{
identified cutoffs
}
\description{
Analyzes row-wise histograms by selecting the global maximum (0),
as well as one further local maximum right (+) and left (-)
}
