% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxValues.R
\name{getTxValues}
\alias{getTxValues}
\title{Get TX CNV values}
\usage{
getTxValues(data, ctrl, ctrlAll, tx, output = "diff",
    statistic = "wilcoxon", op = "median")
}
\arguments{
\item{data}{input data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{tx}{vector of list of transcripts#'}

\item{output}{"ratio" (Sample/Ctrl) or "diff" (Sample-Ctrl)}

\item{statistic}{statistics test to compare groups (controls
vs smp); "t.test" or "wilcoxon" (Mann-Whitney-U Test)}

\item{op}{median or mean}
}
\value{
list with tx-names, p-values and used statistical test
}
\description{
Calculates CN values for given list of transcripts.
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

