% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxValues.R
\name{getTxValuesFast}
\alias{getTxValuesFast}
\title{Get TX CNV values}
\usage{
getTxValuesFast(data, ctrl, ctrlAll, tx)
}
\arguments{
\item{data}{input data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{tx}{vector of list of transcripts#'}
}
\value{
list with tx-names, p-values and used statistical test
}
\description{
Calculates CN values for given list of transcripts, uses standard
values to increase speed: median; wilcoxon; difference
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

