% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{segmentData}
\alias{segmentData}
\title{Binarize Data}
\usage{
segmentData(data, cutoffs, effectsize = c(0, 0))
}
\arguments{
\item{data}{inputdata}

\item{cutoffs}{cutoffs, calculted e.g. with \code{findCutoffs()}}

\item{effectsize}{minimal required effectsize as vector (loss, gain)}
}
\value{
binarized values
}
\description{
Binarizes Data (-1, 0, +1) according to given cutoff list
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

