% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentDataAbs.R
\name{segmentDataAbs}
\alias{segmentDataAbs}
\title{Define gain / loss with given cutoffs}
\usage{
segmentDataAbs(data, upper, lower, ylim = NULL, plot = TRUE)
}
\arguments{
\item{data}{segmentMatrix, e.g. acquired with createSegmentMatrix}

\item{upper}{Upper cutoff (+1)}

\item{lower}{Lower cutoff (-1)}

\item{ylim}{y-axis limits of cutoff plot}

\item{plot}{Should the data be plotted? Might take a long time.}
}
\value{
matrix, containing -1,0,+1 for losses / gains
}
\description{
Use absolute cutoffs to create CNV gain / loss matrix.
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

