% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBins.R
\name{createBins}
\alias{createBins}
\title{Bin CN data}
\usage{
createBins(data, ctrl, ctrlAll, statistic = "wilcoxon", binsize = 2e+05,
    output = "diff")
}
\arguments{
\item{data}{sample data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{statistic}{statistics test to compare groups (controls 
vs smp); "t.test" or "wilcoxon" (Mann-Whitney-U Test)}

\item{binsize}{binsize}

\item{output}{"ratio" (Sample/Ctrl) or "diff" (Sample-Ctrl)}
}
\value{
bins with their corresponding value
}
\description{
Calculate mean, median, sd for each bin for a given binsize.
}
\examples{
print("Please refer to the 'completeWorkflow' vignette!")
}

