% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBins.R
\name{createBins}
\alias{createBins}
\title{Bin CN data}
\usage{
createBins(data, ctrl, ctrlAll, statistic = "wilcoxon", binsize = 2e+05,
  output = "diff", arrayType = "auto", noCores = -1)
}
\arguments{
\item{data}{sample data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{statistic}{statistics test to compare groups (controls 
vs smp); "t.test" or "wilcoxon" (Mann-Whitney-U Test)}

\item{binsize}{binsize}

\item{output}{"ratio" (Sample/Ctrl) or "diff" (Sample-Ctrl)}

\item{arrayType}{"auto","450k", "EPIC"; auto -> tries to automatically 
determine the array type (450k, EPIC)}

\item{noCores}{number of cores for parallelization: needed 
to pass the cran examples :/}
}
\value{
bins with their corresponding value
}
\description{
Calculate mean, median, sd for each bin for a given binsize.
}
\examples{
data <- minfi::getCN(minfi::preprocessRaw(minfiData::RGsetEx))
ctrlAll <- data[,4,drop=FALSE]
ctrl <- data[,5] # apply(ctrlAll, 1, "median")
data <- data[,1,drop=FALSE]
createBins(data,ctrl,ctrlAll, binsize=50000000, noCores=2)[1:3,]
}
