% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConumeeMatrix.R
\name{createConumeeMatrix}
\alias{createConumeeMatrix}
\title{Create matrix from conumee data}
\usage{
createConumeeMatrix(data)
}
\arguments{
\item{data}{output of runConumee(), might contain segments,
bins or transcripts}
}
\value{
n x m matrix, n beeing the amoutn of segments / bins;
m equals the amount of samples
}
\description{
Create matrix from conumee segment / bin data.
}
\examples{
require(IlluminaHumanMethylation450kanno.ilmn12.hg19)
norm <- minfi::preprocessRaw(minfiData::RGsetEx)
samples <- norm[,1:3]
ctrl <- norm[,4:6]
candidatesDATA <- runConumee(samples,ctrl,"segments")
createConumeeMatrix(candidatesDATA)
}

