% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasenMset.R
\name{dasenMset}
\alias{dasenMset}
\title{Adapted dasen function}
\usage{
dasenMset(mns, uns, onetwo)
}
\arguments{
\item{mns}{getMethyl()}

\item{uns}{getUnmethyl()}

\item{onetwo}{probe Types ("I" or "II")}
}
\value{
dasen normalized MethylSet
}
\description{
dasen function from the wateRmelon package, 
returns, however, a MethylSet. Code from bgeqot.R.
}
\examples{
m <- minfi::getMeth(minfiData::MsetEx[,1])
u <- minfi::getUnmeth(minfiData::MsetEx[,1])
#ot <- minfi::getAnnotation(minfiData::MsetEx)$Type
ot <- c(rep("I",200), rep("II", (length(u[,1])-200)))
dasenMset(m,u,ot)
}
