% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSegments.R
\name{findSegments}
\alias{findSegments}
\title{Find Segments in the provided CN data}
\usage{
findSegments(data, ctrl, ctrlAll, statistic = "wilcoxon", plot = FALSE,
  delta = 500, output = "diff", ylim = NULL, arrayType = "auto")
}
\arguments{
\item{data}{CN data to evaluate}

\item{ctrl}{CN data of controls, levels to test to (1 mean / median
over all ctrl samples)}

\item{ctrlAll}{CN data of all control samples}

\item{statistic}{statistics test to compare groups (controls vs smp);
"t.test" or "wilcoxon" (Mann-Whitney-U Test)}

\item{plot}{plot changepoints}

\item{delta}{area around changepoints to plot for plot=T}

\item{output}{"ratio" (Sample/Ctrl) or "diff" (Sample-Ctrl)}

\item{ylim}{if plot=T give ylim}

\item{arrayType}{"auto","450k", "EPIC"; auto -> tries to automatically 
determine the array type (450k, EPIC)}
}
\value{
data containing chr, startCG, endCG, segmentmedian, -mean, 
p-value, SD and samplename
}
\description{
Uses data from minfis getCN() function and normalizes 
probe-wise against control CN data.
Segments are identified with changepoints cpr.var() 
function (BinSeg)
}
\examples{
norm <- minfi::getCN(minfi::preprocessRaw(minfiData::RGsetEx))
ctrlAll <- norm[,5,drop=FALSE]
ctrl <- norm[,4] #ctrl <- apply(ctrlAll, 1, "median")
samples <- norm[,1,drop=FALSE]
findSegments(samples,ctrl, ctrlAll)[1:4,]
}

