% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxValues.R
\name{getTxValuesFast}
\alias{getTxValuesFast}
\title{Get TX CNV values}
\usage{
getTxValuesFast(data, ctrl, ctrlAll, tx)
}
\arguments{
\item{data}{input data}

\item{ctrl}{control data}

\item{ctrlAll}{CN data of all control samples}

\item{tx}{vector of list of transcripts#'}
}
\value{
list with tx-names, p-values and used statistical test
}
\description{
Calculates CN values for given list of transcripts, uses standard
values to increase speed: median; wilcoxon; difference
}
\examples{
#' data <- data.frame(
smp1=c(-8.12, -5.225, -3.24, -3.61),
smp2=c(-5.0, -3.98, -4.06, -4.5),
smp3=c(NA, -2.48, -2.27, -2.1)
)
ctrlAll <- data.frame(
ctl1=c(1.0, -3.6, 0.7, -0.73),
ctl2=c(-0.4, -4.1, -4.2, -3.9),
ctl2=c(0.74, -1.12, -2.8, -1.67)
)
rownames(data) <- c(
"cg05132306", "cg15527168",
"cg17434257", "cg17592667"
)
rownames(ctrlAll) <- rownames(data)
ctrl <- apply(ctrlAll, 1, "median")
names(ctrl) <- rownames(data)
getTxValues(data, ctrl, ctrlAll, "uc001aih.1")
}

