% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runConumee.R
\name{runConumee}
\alias{runConumee}
\title{Analyze data with conumee}
\usage{
runConumee(data, ctrl, what = "segments", tx = NULL)
}
\arguments{
\item{data}{MethylSet: sample data}

\item{ctrl}{MethylSet: reference}

\item{what}{type of return data: segments, bins, transcripts}

\item{tx}{vector of transcripts of interests}
}
\value{
list with $data and $what, containing the CNV.write conumee outputs
}
\description{
Analyzes the given 450k data with conumee, extracting 
either segments, bins or transcripts.
}
\examples{
print("A runnable example is also shown in the 
createConumeeMatrix examples")
#require(IlluminaHumanMethylation450kanno.ilmn12.hg19)
#norm <- minfi::preprocessRaw(minfiData::RGsetEx)
#data <- norm[,1]
#ctrl <- norm[,4]
#runConumee(data,ctrl,"segments")
}
